handles #197

+++ src/astroprint/camera/v4l2/gstreamer/pipeline.py
@@ -75,7 +75,7 @@ class AstroPrintPipeline(object):
 		if self._process:
 			self._sendReqToProcess({'action': 'shutdown'})
 			self._process.join(2.0) #Give it two seconds to exit and kill otherwise
-			if self._process.exitcode is None:
+			if self._process and self._process.exitcode is None:
 				self._logger.warn('Process did not shutdown properly. Terminating...')
 				self._process.terminate()
 				self._process.join(2.0) # Give it another two secods to terminate, otherwise kill
